+++ django_project/geohosting/src/pages/DashboardPage/DashboardPage.tsx
@@ -12,11 +12,11 @@ import DashboardSidePanel
   from "../../components/DashboardSidePanel/DashboardSidePanel";
 import { Route, Routes, useLocation } from "react-router-dom";
 import DashboardMainPage from "./DashboardMainPage";
-import SupportPage from "./Support/SupportPage";
 import OrdersList from './Orders/OrderList';
 import OrderDetail from "./Orders/OrderDetail";
 import ProfilePage from './Profile/ProfilePage';
 import AgreementPage from './AgreementPage';
+import SupportList from "./Support/SupportList";
 
 const DashboardPage = ({ title = "Dashboard" }) => {
   const [isOpen, setIsOpen] = React.useState(false);
@@ -61,7 +61,7 @@ const DashboardPage = ({ title = "Dashboard" }) => {
             <Routes>
               <Route path="/" element={<DashboardMainPage/>}/>
               <Route path='/agreements' element={<AgreementPage/>}/>
-              <Route path="/support" element={<SupportPage/>}/>
+              <Route path="/support" element={<SupportList/>}/>
               <Route path="/orders/:id" element={<OrderDetail/>}/>
               <Route path="/orders" element={<OrdersList/>}/>
               <Route path='/profile' element={<ProfilePage/>}/>
