+++ django_project/geohosting/src/redux/types/paginationTypes.tsx
@@ -1,0 +1,6 @@
+export interface PaginationResult {
+  count: number,
+  next: string | null,
+  previous: string | null,
+  results: any[]
+}
